services/ussd_service/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
django-cors-headers>=4.3.1
gunicorn>=21.2.0
django-environ
